refactor : format 함수 추출
